# perf: reuse typeof result