@SukkaW 2021-11-16
Reuse `y` sounds like a bad idea. However, the type (`string`) doesn't change so I guess v8 could handle this.